chore: log 제거
