+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import kotlinx.coroutines.launch
-import woowacourse.shopping.BaseViewModelFactory2
+import woowacourse.shopping.BaseViewModelFactory
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.data.ProductRepository
@@ -46,7 +46,7 @@ class MainViewModel(
             viewModelFactory {
                 initializer {
                     val appContainer = (this[APPLICATION_KEY] as ShoppingApplication).container
-                    BaseViewModelFactory2(appContainer).create(MainViewModel::class.java)
+                    BaseViewModelFactory(appContainer).create(MainViewModel::class.java)
                 }
             }
     }
